Recoded
Basically rewrote this from the ground up. switched to arrays and a function that would just  pull in values and compare them. 

Still haven't solved how to detect larger values so it can tell the difference between IV and VI. 
I feel like I'm so close I could trip on it but I just can't seem to get exactly the right vocabulary for it out.